added icons to toolbar and deleted edit and delete buttons from speechview
