+++ app/src/main/java/com/google/cloud/android/speech/ScriptView.java
@@ -6,6 +6,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.text.method.ScrollingMovementMethod;
 import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -46,6 +48,14 @@ public class ScriptView extends AppCompatActivity {
         }
     }
 
+    public void goToEditSpeech(View view) {
+        Intent intent = new Intent(this, NewSpeech.class);
+
+        intent.putExtra("filename", speechName);
+        intent.putExtra("scriptText", scriptText);
+        startActivity(intent);
+    }
+
     private void setScriptText() {
         // Get text body
         TextView scriptBody = (TextView) findViewById(R.id.scriptBody);
@@ -59,7 +69,24 @@ public class ScriptView extends AppCompatActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.base_menu, menu);
+        getMenuInflater().inflate(R.menu.script_view_menu, menu);
         return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_edit) {
+            View view = findViewById(R.id.action_edit);
+            goToEditSpeech(view);
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
     }
 }
