+++ app/src/main/res/menu/script_view_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".MainActivity">
+
+    <item
+        android:id="@+id/action_edit"
+        android:orderInCategory="300"
+        android:title="User"
+        android:icon="@drawable/ic_pencil"
+        app:showAsAction="ifRoom" >
+    </item>
+
+</menu>
