+++ app/src/main/res/menu/speech_view_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".MainActivity">
+
+    <item
+        android:id="@+id/action_delete"
+        android:orderInCategory="300"
+        android:title="User"
+        android:icon="@drawable/ic_delete"
+        app:showAsAction="ifRoom" >
+    </item>
+
+</menu>
